requirements.txt
configparser==3.5.0
coreapi==2.1.1
Django==1.10.4
django-filter==1.0.1
django-rest-swagger==2.1.1
django-webpack-loader==0.4.1
djangorestframework==3.5.3
enum34==1.1.6
flake8==3.2.1
futures==3.0.5
gevent==1.2.0
greenlet==0.4.11
gunicorn==19.6.0
itypes==1.1.0
json-logging-py==0.2
mccabe==0.5.3
openapi-codec==1.2.1
psycopg2==2.6.2
pycodestyle==2.2.0
pyflakes==1.3.0
simplejson==3.10.0
uritemplate==3.0.0
